requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu130
dvc==3.66.1
hydra-core==1.3.2
mlflow==3.10.1
xarray==2025.6.1
netCDF4==1.7.4
h5netcdf==1.8.1
pyyaml==6.0.3
omegaconf==2.3.0
torch==2.10.0
torchvision==0.25.0
dask==2026.3.0
apache-airflow==3.1.8
zenodo_get==3.0.0
plotly==6.6.0